requirements.txt
attrs==20.2.0
backcall==0.2.0
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coloredlogs==14.0
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==2.1.1
django-click==2.2.0
django-rq==2.3.2
django-s3-sqlite==0.0.3
djangorestframework==3.11.1
gitdb==4.0.5
GitPython==3.1.8
humanfriendly==8.2
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
inflection==0.5.1
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
json-delta==2.0
kiwisolver==1.2.0
lxml==4.5.2
matplotlib==3.3.2
<<<<<<< HEAD
=======
mysqlclient==1.3.12
>>>>>>> e281625346f12ddae273a486a5ef4fea1ca65dda
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
parso==0.7.1
pbr==5.5.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
Pygments==2.7.1
pyparsing==2.4.7
pyreadline==2.1
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
redis==3.5.3
regparser==4.3.1
requests==2.24.0
requests-cache==0.5.2
roman==3.3
rq==1.5.2
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore==3.2.2
tifffile==2020.9.3
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.2.0
